m4v_h263_enc: Use uint32_t for ULong
ULong is used in the encoder code as 32-bit unsigned integer type.
But it has been defined as an unsigned long, which will be 64-bit
on 64-bit platforms. This causes many problems including output
mismatch between 32-bit and 64-bit code and crash in certain cases.

This patch defines ULong as uint32_t. In addition, it uses uintptr_t
at places where an address is cast to an integer to check for alignment.

Change-Id: I2b0086824a5985b9fc4628ce6fbe81db156e37d2
Signed-off-by: Ashok Bhat <ashok.bhat@arm.com>